Satyam

+++ app/Http/Controllers/Master/DepartmentController.php
@@ -33,13 +33,10 @@ class DepartmentController extends Controller
     public function forceDelete($id)
     {
         $departments = Department::withTrashed()->find($id);
-
         if (!$departments) {
             return response()->json(['success' => false], 404);
         }
-
         $departments->forceDelete();
-
         return response()->json(['success' => true]);
     }
 
@@ -57,21 +54,17 @@ class DepartmentController extends Controller
                 'min:2',
                 Rule::notIn(['']),
             ],
+            'unique' => 'required|unique:departments,name,except,id'
         ], [
             'name.regex' => 'The :attribute may only contain letters and spaces. Numbers and special characters are not allowed.',
         ]);
-
-        // Create the department using the Department model
+        // dd($request->all());
         Department::create([
             'name' => $request->name,
-            // Add other fields if needed
+            'unique_id'=>$request->unique,
         ]);
-
-        // Redirect back to the list of departments
         return redirect('/departments')->with('message', 'Department created successfully.');
     }
-
-    // Show the list of departments
     public function index()
     {
         $departments = Department::all();
@@ -98,7 +91,7 @@ class DepartmentController extends Controller
                 'regex:/^[A-Za-z]+( [A-Za-z]+)*$/',
                 'min:2',
                 Rule::notIn(['']),
-            ],
+            ],'unique' => 'required|unique:departments,name,except,id'
         ], [
             'name.regex' => 'The :attribute may only contain letters and spaces. Numbers and special characters are not allowed.',
         ]);
@@ -107,10 +100,8 @@ class DepartmentController extends Controller
         $department = Department::findOrFail($id);
         $department->update([
             'name' => $request->name,
-            // Add other fields if needed
+            'unique_id'=>$request->unique,
         ]);
-
-        // Redirect back to the list of departments
         return redirect('departments')->with('message', 'Department updated Successfully!');
     }
 
@@ -118,11 +109,14 @@ class DepartmentController extends Controller
     public function destroy($id)
     {
         // Find the department and delete it
-        $department = Department::findOrFail($id);
+        $department = Department::find($id);
+        $referencesExist = $department->users()->exists() || $department->designations()->exists();
+        // dd($referencesExist);
+        if ($referencesExist) {
+            return response()->json(['success' => false, 'message' => 'Department is referenced in one or more tables and cannot be deleted.']);
+        }
         $department->delete();
-
-        // Redirect back to the list of departments
-        return response()->json(['success' => true]);
+        return response()->json(['success' => true, 'message' => 'Department deleted successfully.']);
     }
     public function updateStatus(Request $request, Department $department)
     {
@@ -145,7 +139,6 @@ class DepartmentController extends Controller
     }
     public function import(Request $request){
         Excel::import(new DepartmentImport, $request->file('file'));
-
         return redirect()->back();
     }
 }
